% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handcode.R
\name{button_output}
\alias{button_output}
\title{button_output: Format output for `shiny::radioButtons`}
\usage{
button_output(classification, button, names = FALSE)
}
\arguments{
\item{classification}{List; a classification item obtained from `data_for_app()`.}

\item{button}{Numeric; indicates for which button the names and labels should be returned.}

\item{names}{Logical; if TRUE, the function returns a formatted list of labels; if FALSE, it returns a vector of names suitable for `shiny::radioButtons()`.}
}
\value{
A list of HTML-formatted labels or a vector of category names for `shiny::radioButtons()` in the Shiny app, depending on the value of the `names` parameter.
}
\description{
`button_output()` is an internal function used by `handcode()`. It operates within the Shiny app `handcoder_app()` to format the appearance of `shiny::radioButtons()`.
}
