% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handcode.R
\name{gen_output}
\alias{gen_output}
\title{gen_output: Combine data and values from `handcoder_app` to generate output}
\usage{
gen_output(data, values)
}
\arguments{
\item{data}{Data frame; the dataset initially passed to `handcoder_app`.}

\item{values}{Values generated by the server of `handcoder_app`.}
}
\value{
A data frame containing the coded variables.
}
\description{
`gen_output` is an internal function used by `handcode`. It operates within the Shiny app `handcoder_app` to prepare the output data frame.
}
