% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handcode.R
\name{handcoder_app}
\alias{handcoder_app}
\title{handcoder_app: Shiny app to classify text snippets}
\usage{
handcoder_app(a)
}
\arguments{
\item{a}{List; contains all information required for the app to function. This list is automatically generated by `data_for_app`.}
}
\value{
A data frame containing the coded variables produced by the Shiny app.
}
\description{
`handcoder_app` is an internal function used by `handcode`. It launches a Shiny app that allows users to review text snippets and classify them manually according to predefined categories.
}
