% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handcode.R
\name{handcoder_ui}
\alias{handcoder_ui}
\title{handcodeer_ui: UI for the internal `handcoder_app()` application}
\usage{
handcoder_ui(a, mode = "r")
}
\arguments{
\item{a}{List; contains all data and configuration settings required by the annotation app. This list is generated by `data_for_app()` and includes the text to annotate, annotation variables, and other runtime parameters.}

\item{mode}{Character; either `"r"` (interactive R session) or `"web"` (running inside a parent Shiny app). Determines how the app exits and returns data.}
}
\value{
A Shiny UI object to be passed to `shiny::shinyApp()` together with `handcoder_server()`.

Shiny UI
}
\description{
`handcoder_ui()` is an internal function used by `handcode()`. It creates the Shiny user interface of the interactive annotation app created by `handcoder_app()`.
}
\details{
This function is not intended to be called directly by users. It is automatically used when launching the interactive annotation workflow.
}
