% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handcode.R
\name{button_output}
\alias{button_output}
\title{button_output: Formatting of output for shiny::radioButtons}
\usage{
button_output(classification, button, names = FALSE)
}
\arguments{
\item{classification}{Classification list obtained as list item from `data_for_app`}

\item{button}{Numerical value indicating for which button names and labels should be returned}

\item{names}{Logical value indicating whether the function is supposed to return a formatted list of labels or a vector of names for use in radioButtons.}
}
\value{
Depending on the logical names input, the function returns a list of HTML formatted labels or the list of category names for the display of the radioButtons in the Shiny-App.
}
\description{
`button_output` is an internal function to `handcode`. It works within the Shiny-App `handcoder_app` to format the look of the shiny::radioButtons.
}
