% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handcode.R
\name{character_to_data}
\alias{character_to_data}
\title{character_to_data: Transform text vector and arg_list to dataframe.}
\usage{
character_to_data(data, arg_list)
}
\description{
character_to_data: Transform text vector and arg_list to dataframe.
}
