% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handcode.R
\name{data_for_app}
\alias{data_for_app}
\title{data_for_app: Prepare inputs for the Shiny app}
\usage{
data_for_app(
  data,
  start,
  randomize,
  context,
  pre = NULL,
  post = NULL,
  comparison_pre = NULL,
  comparison_post = NULL
)
}
\arguments{
\item{data}{Data frame; prepared by `character_to_data()` or previously returned by `handcode()`.}

\item{start}{Numeric; the start value provided to `handcode()`.}

\item{randomize}{Logical; the randomize value provided to `handcode()`.}

\item{context}{Logical; the context value provided to `handcode()`.}

\item{pre}{Optional character vector; texts that appear before each corresponding text to be coded.}

\item{post}{Optional character vector; texts that appear after each corresponding text to be coded.}

\item{comparison_pre}{Optional character vector; texts that appear before each corresponding comparison text to be coded.}

\item{comparison_post}{Optional character vector; texts that appear after each corresponding comparison text to be coded.}
}
\value{
A list of inputs required by the Shiny app.
}
\description{
`data_for_app()` is an internal function used by `handcode()`. It takes the inputs `data`, `start`, `randomize`, and `context` to generate a list of inputs that can be passed to `handcoder_app()`.
}
