% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handcode.R
\name{gen_output}
\alias{gen_output}
\title{gen_output: Combine data and values of shiny app to output}
\usage{
gen_output(data, values)
}
\arguments{
\item{data}{The dataframe initially passed to `handcoder_app` within a.}

\item{values}{The values generated by the server of the `handcoder_app`.}
}
\description{
`gen_output` is an internal function to `handcode`. It works within the shiny app `handcoder_app` to prepare the output dataframe.
}
