% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handcode.R
\name{gen_output}
\alias{gen_output}
\title{gen_output: Combine data and values of shiny app to output}
\usage{
gen_output(data, values)
}
\description{
gen_output: Combine data and values of shiny app to output
}
