% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handcode.R
\name{handcode}
\alias{handcode}
\title{handcode: Classifying text into pre-defined categories.}
\usage{
handcode(data, start = "first_empty")
}
\arguments{
\item{data}{A dataframe initialized via the init_data() function. If you want to continue an ongoing handcoding process, you can also load your most recent handcode() output.}

\item{start}{A numeric value indicating the line in which you want to start handcoding. Alternatively, you can set start to "first_empty" to automatically start handcoding in the first line that has not been handcoded yet.}
}
\description{
`handcode` opens a Shiny app which allows for handcoding strings of text into pre-defined categories. You can code between one and three variables at a time. It returns an updated data.frame with your handcoded classifications.
}
