% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handcode.R
\name{handcode}
\alias{handcode}
\title{handcode: Classifying text into pre-defined categories.}
\usage{
handcode(data, start = 1, newname = "data_new")
}
\arguments{
\item{data}{A dataframe initialized via the init_data() function.}

\item{start}{A numeric value indicating the line in which you want to start handcoding. Alternatively, you can set start to "first_empty" to automatically start handcoding in the first line that has not been handcoded yet.}

\item{newname}{A character string indicating a name for a new object that saves the handcoding to your global environment}
}
\description{
`handcode` opens a Shiny app which allows for handcoding of texts into pre-defined categories. The results will be saved as a new object to the global environment.
}
