% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handcodeR-package.R
\name{handcodeR-package}
\alias{handcodeR-package}
\alias{handcodeR}
\title{handcodeR: Text Annotation App}
\description{
Shiny-based application for hand-coding text data into predefined
categories. Supports categorial multi-class annotation, binary
two-choice annotation, paired-text comparison workflows, and recovery
via autosave and quicksave.
}
\section{Main functions}{

\itemize{
  \item \code{\link{handcode}}: categorial multi-class annotation
  \item \code{\link{handcode_binary}}: binary / two-choice annotation
}
}

\author{
Maintainer: Lukas Isermann \email{lukas.isermann@uni-mannheim.de}
(\href{https://orcid.org/0000-0002-7195-9302}{ORCID})

Co-Author: Dennis Klingenspohr \email{dennis.klingenspohr@gmail.com}
(\href{https://orcid.org/0009-0000-7876-5634}{ORCID})
}
\keyword{internal}
