% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handcode.R
\name{handcoder_app}
\alias{handcoder_app}
\title{handcoder_app: Classifying text into pre-defined categories.}
\usage{
handcoder_app(a)
}
\description{
handcoder_app: Classifying text into pre-defined categories.
}
