% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handcode.R
\name{handcoder_app}
\alias{handcoder_app}
\title{handcoder_app: App to classify text snippets}
\usage{
handcoder_app(a)
}
\arguments{
\item{a}{A list containing all relevant information for the app to function. The list will be automatically generated by `data_for_app`.}
}
\value{
The shinyApp returns a dataframe with the coded variables.
}
\description{
`handcoder_app` is an internal function to `handcode`. It provides a shiny app that allows users to click through different snippets of text and classify them manually according to predefined categories.
}
