% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_data.R
\name{init_data}
\alias{init_data}
\title{init_data: Initialize dataframe for handcode().}
\usage{
init_data(texts, categories)
}
\arguments{
\item{texts}{A vector of character strings you want to classify by hand.}

\item{categories}{A vector of character strings giving names for different categories you want to classify.}
}
\description{
`init_data` initializes a dataframe to be passed to handcode(). The function takes a vector of texts and a number of categories by which you want to classify the texts as input and returns a dataframe that can be used as data input in handcode().
}
